Minor overall improvements were made

+++ src/main/java/com/example/elective/dto/CompletedCourseDTO.java
@@ -36,7 +36,7 @@ public class CompletedCourseDTO {
     return courseDTO.endDate;
   }
 
-  public int getDuration() {
+  public long getDuration() {
     return courseDTO.getDuration();
   }
